apps/worker/requirements.txt
openai-whisper==20231117
google-generativeai==0.8.3
minio==7.2.15
redis==5.2.1
pydantic==2.10.6
python-dotenv==1.0.1
